# applications/v4/api/openapi.yaml
openapi: 3.0.3
info:
  title: Weekly Planning API
  version: 1.0.0
  description: API contract for accounts, persons, projects, weeks, and weekly planning.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      x-spring-paginated: true
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/SizeQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
        - $ref: '#/components/parameters/SearchQueryParam'
        - $ref: '#/components/parameters/ProjectStatusQueryParam'
      responses:
        '200': { $ref: '#/components/responses/ProjectList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      tags: [Projects]
      operationId: createProject
      requestBody:
        $ref: '#/components/requestBodies/CreateProjectRequestBody'
      responses:
        '201': { $ref: '#/components/responses/ProjectCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /projects/{projectId}:
    get:
      tags: [Projects]
      operationId: getProjectById
      parameters:
        - $ref: '#/components/parameters/ProjectIdPathParam'
      responses:
        '200': { $ref: '#/components/responses/ProjectDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      tags: [Projects]
      operationId: updateProjectById
      parameters:
        - $ref: '#/components/parameters/ProjectIdPathParam'
      requestBody:
        $ref: '#/components/requestBodies/UpdateProjectRequestBody'
      responses:
        '200': { $ref: '#/components/responses/ProjectDetails' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Projects]
      operationId: deleteProjectById
      parameters:
        - $ref: '#/components/parameters/ProjectIdPathParam'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    LocationHeader:
      description: URI of the created resource.
      schema:
        type: string
        format: uri
  parameters:
...
    ProjectIdPathParam:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PageQueryParam:
      name: page
      in: query
      schema: { type: integer, format: int32, minimum: 0, default: 0 }
    SizeQueryParam:
      name: size
      in: query
      schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 20 }
    SortQueryParam:
      name: sort
      in: query
      schema:
        type: array
        items: { type: string }
      style: form
      explode: true
    SearchQueryParam:
      name: search
      in: query
      schema: { type: string, minLength: 1, maxLength: 120 }
...
    ProjectStatusQueryParam:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/ProjectStatus' }
...
  requestBodies:
...
    CreateProjectRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CreateProjectRequest' }
    UpdateProjectRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UpdateProjectRequest' }
...
  responses:
    NoContent:
      description: No content.
...
    ProjectList:
      description: Paginated projects.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProjectPage' }
    ProjectDetails:
      description: Project details.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProjectDetail' }
    ProjectCreated:
      description: Project created.
      headers:
        Location: { $ref: '#/components/headers/LocationHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProjectDetail' }
...
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ProjectStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    SortDirection:
      type: string
      enum: [asc, desc]
...
    CreateProjectRequest:
      type: object
      required: [name, code, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        code:
          type: string
          minLength: 1
          maxLength: 40
        status:
          $ref: '#/components/schemas/ProjectStatus'
    UpdateProjectRequest:
      type: object
      required: [name, code, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        code:
          type: string
          minLength: 1
          maxLength: 40
        status:
          $ref: '#/components/schemas/ProjectStatus'
    ProjectSummary:
      type: object
      required: [projectId, name, code, status]
      properties:
        projectId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        code:
          type: string
          maxLength: 40
        status:
          $ref: '#/components/schemas/ProjectStatus'
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/ProjectSummary'
        - type: object
          required: [createdAt, updatedAt]
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    ProjectPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProjectSummary'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortOrder'
    SortOrder:
      type: object
      required: [property, direction]
      properties:
        property:
          type: string
        direction:
          $ref: '#/components/schemas/SortDirection'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [timestamp, status, error, code, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        code:
          type: string
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'